Refactor race details form in separate component
